+++ Matrix/pzskylmat.h
@@ -199,8 +199,8 @@ public:
   int Decompose_Cholesky_blk(long blk_sz);
 
   int Decompose_LDLt    ();  // Faz A = LDLt.
-  int Decompose_Cholesky(std::list<int> &singular);  // Faz A = GGt.
-  int Decompose_LDLt    (std::list<int> &singular);  // Faz A = LDLt.
+  int Decompose_Cholesky(std::list<long> &singular);  // Faz A = GGt.
+  int Decompose_LDLt    (std::list<long> &singular);  // Faz A = LDLt.
   
   int Subst_Forward  ( TPZFMatrix<TVar> *b ) const;
   int Subst_Backward ( TPZFMatrix<TVar> *b ) const;
